requirements.txt
Flask==2.3.2
gunicorn==21.2.0
fpdf==1.7.2
qrcode==7.4.2
Pillow==10.0.0
google-api-python-client==2.123.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
